libcamera: v4l2_videodevice: Reduce zero sequence warning log level
The V4L2VideoDevice ensures that all sequence numbers for applications
commence at zero from the libcamera perspective. This should be the
behaviour expected by kernel drivers, but this is not always the case.

This is handled internally to ensure consistency, and a warning is
printed if the device does not start from zero. It was expected that the
Warning would help highlight where kernel drivers should be fixed, but
this has led to several false positive reports of failures where people
have been concerned that this warning is a cause for unrelated issues.

Lower the log level print to 'Info', to reduce the apparent severity of
this warning. Info is likely more appropriate that Debug to continue to
facilitate awareness of kernel drivers that could be improved, while not
appearing to be a fault.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>