libcamera: Add todo comment in meson.build about switch to dependency('dl')
The dlopen() & co. functions are provided by libdl on some systems, and
but the C library on others. Starting from version 0.62.0, meson handles
this behind the scenes when using dependency('dl'). Add a todo comment
to remember we should replace the manual implementation with a
dependency() call when updating to meson 0.62.0 or newer.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>